Add test for zipstream retrieving from s3
