+++ tests/Feature/MediaStreamTest.php
@@ -76,25 +76,4 @@ class MediaStreamTest extends TestCase
 
         $this->assertEquals(2, $zipStreamResponse->getMediaItems()->count());
     }
-
-    protected function assertFileExistsInZip($zipPath, $filename)
-    {
-        $this->assertTrue($this->fileExistsInZip($zipPath, $filename), "Failed to assert that {$zipPath} contains a file name {$filename}");
-    }
-
-    protected function assertFileDoesntExistsInZip($zipPath, $filename)
-    {
-        $this->assertFalse($this->fileExistsInZip($zipPath, $filename), "Failed to assert that {$zipPath} doesn't contain a file name {$filename}");
-    }
-
-    protected function fileExistsInZip($zipPath, $filename): bool
-    {
-        $zip = new ZipArchive();
-
-        if ($zip->open($zipPath) === true) {
-            return $zip->locateName($filename, ZipArchive::FL_NODIR) !== false;
-        }
-
-        return false;
-    }
 }
