+++ tests/Feature/S3Integration/S3IntegrationTest.php
@@ -5,9 +5,14 @@ namespace Spatie\MediaLibrary\Tests\Feature\S3Integration;
 use Carbon\Carbon;
 use Aws\S3\S3Client;
 use Illuminate\Support\Facades\Artisan;
+use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Storage;
+use Spatie\MediaLibrary\MediaStream;
+use Spatie\MediaLibrary\Models\Media;
 use Spatie\MediaLibrary\Tests\TestCase;
 use Illuminate\Contracts\Filesystem\Factory;
+use Spatie\TemporaryDirectory\TemporaryDirectory;
+use Symfony\Component\HttpFoundation\StreamedResponse;
 
 class S3IntegrationTest extends TestCase
 {
@@ -277,6 +282,26 @@ class S3IntegrationTest extends TestCase
         $this->assertS3FileNotExists($derivedMissingImageOriginal);
         $this->assertSame($existsCreatedAt, Storage::disk('s3_disk')->lastModified($derivedImageExists));
         $this->assertGreaterThan($missingCreatedAt, Storage::disk('s3_disk')->lastModified($derivedMissingImage));
+    }
+
+    /** @test */
+    public function it_can_retrieve_a_zip_with_s3_disk()
+    {
+        $media = $this->testModel
+            ->addMedia($this->getTestJpg())
+            ->toMediaCollection('default', 's3_disk');
+
+        $zipStreamResponse = MediaStream::create('my-media.zip')->addMedia($media);
+
+        ob_start();
+        @$zipStreamResponse->toResponse(request())->sendContent();
+        $content = ob_get_contents();
+        ob_end_clean();
+
+        $temporaryDirectory = (new TemporaryDirectory())->create();
+        file_put_contents($temporaryDirectory->path('response.zip'), $content);
+
+        $this->assertFileExistsInZip($temporaryDirectory->path('response.zip'), 'test.jpg');
     }
 
     protected function cleanUpS3()
